[BGM-6] Rename 'News' entity to 'NewsItem' [BGM-5] Rename 'Events' enity class to Event [BGM-4] Rename 'entities' package to entity [BGM-3] Remove Main.java [BGM-2] Rename DAO package to lowercase [BGM-1] Rename the packages corresponding to the project
