+++ src/by/bgminsk/dao/UserDao.java
@@ -1,0 +1,7 @@
+package by.bgminsk.dao;
+
+import by.bgminsk.entity.User;
+
+public interface UserDao{
+    void save(User user);
+}
