+++ src/by/bgminsk/entity/NewsItem.java
@@ -1,15 +1,14 @@
-package com.company.entities;
+package by.bgminsk.entity;
 
-import java.util.Date;
 import java.util.Objects;
 
-public class News {
+public class NewsItem {
     private int id;
     private String title;
     private String content;
     private String author;
 
-    public News(int id, String title, String content, String author) {
+    public NewsItem(int id, String title, String content, String author) {
         this.id = id;
         this.title = title;
         this.content = content;
@@ -52,8 +51,8 @@ public class News {
     public boolean equals(Object o) {
         if (this == o) return true;
         if (o == null || getClass() != o.getClass()) return false;
-        News news = (News) o;
-        return id == news.id && Objects.equals(title, news.title) && Objects.equals(content, news.content) && Objects.equals(author, news.author);
+        NewsItem newsItem = (NewsItem) o;
+        return id == newsItem.id && Objects.equals(title, newsItem.title) && Objects.equals(content, newsItem.content) && Objects.equals(author, newsItem.author);
     }
 
     @Override
@@ -63,7 +62,7 @@ public class News {
 
     @Override
     public String toString() {
-        return "News{" +
+        return "NewsItem{" +
                 "id=" + id +
                 ", title='" + title + '\'' +
                 ", content='" + content + '\'' +
