--- src/com/company/DAO/UserDao.java
@@ -1,7 +1,0 @@
-package com.company.DAO;
-
-import com.company.entities.User;
-
-public interface UserDao{
-    void save(User user);
-}
